Add timestamped logging
log.p8 contains useful helper functions for logging timestamped
messages to STDOUT. I couldn't find a way to log to STDERR
unfortunately, it doesn't seem that it is supported by the pico8
API. I took a dependency on the tostr function in strings.p8 to handle
stringifying tables, booleans, and other such types.

This commit also updates the README following the pattern there to add
a reference to the new log.p8 file.